requirements.txt
Flask==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.2
requests==2.32.3